Refactor routes and data
