+++ src/routes/health_check.rs
@@ -1,0 +1,5 @@
+use actix_web::HttpResponse;
+
+pub async fn health_check() -> HttpResponse {
+    HttpResponse::Ok().body("OK")
+}
