+++ src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod health_check;
+mod redirect;
+
+pub use health_check::health_check;
+pub use redirect::post_redirect;
